+++ Content/Items/Accessories/Chromagem/Auribloom/AuribloomItem.cs
@@ -6,6 +6,9 @@ using Terraria.DataStructures;
 using Terraria.ID;
 using Terraria.ModLoader;
 using Microsoft.Xna.Framework;
+using System;
+using System.Diagnostics;
+using MonoMod.Cil;
 
 namespace SuperfluityTwo.Content.Items.Accessories.Chromagem.Auribloom
 {
@@ -55,7 +58,7 @@ namespace SuperfluityTwo.Content.Items.Accessories.Chromagem.Auribloom
                 player.AddBuff(ModContent.BuffType<GoldBloodBuff>(), 5 * 60);
             }
 
-            if (!hideVisual && Main.rand.NextBool(8))
+            if ((Main.LocalPlayer == player || (Main.LocalPlayer.team == player.team && player.team != 0)) && !hideVisual && Main.rand.NextBool(8))
             {
                 Dust.NewDustDirect(player.Center + 16 * 7 * Main.rand.NextFloat(MathHelper.TwoPi).ToRotationVector2(), 0, 0, DustID.GoldFlame, Alpha: 0).noGravity = true;
             }
@@ -71,7 +74,8 @@ namespace SuperfluityTwo.Content.Items.Accessories.Chromagem.Auribloom
                 foreach (var plr in Main.ActivePlayers)
                 {
                     if (plr == player) continue;
-                    if (!plr.InOpposingTeam(player) && plr.Center.Distance(player.Center) < 16 * 7 && !plr.dead)
+                    if (plr.team != player.team || player.team == 0) continue;
+                    if (plr.Center.Distance(player.Center) < 16 * 7 && !plr.dead)
                     {
                         plr.AddBuff(ModContent.BuffType<GoldBloodBuff>(), 5 * 60);
                     }
@@ -83,51 +87,82 @@ namespace SuperfluityTwo.Content.Items.Accessories.Chromagem.Auribloom
     public class AuribloomPlayer : ModPlayer
     {
         public bool hasAuri = false;
+        public float auriDelta = 0f;
         public override void ResetEffects()
         {
             hasAuri = false;
         }
-    }
-
-    
-    internal class AuriLayer : PlayerDrawLayer
-	{
-        public static Asset<Texture2D> texAsset;
+        public override void PostUpdateEquips()
+        {
+            if (hasAuri)
+            {
+                auriDelta += 0.1f;
+            }
+            else
+            {
+                auriDelta = 0;
+            }
+        }
 
         public override void Load()
         {
+            IL_Main.DoDraw += DrawAuriZone;
             texAsset = ModContent.Request<Texture2D>($"{nameof(SuperfluityTwo)}/Content/Items/Accessories/Chromagem/Auribloom/AuribloomVisual", AssetRequestMode.AsyncLoad);
         }
+        public override void Unload()
+        {
+            IL_Main.DoDraw -= DrawAuriZone;
+        }
+        
+        
+        public static Asset<Texture2D> texAsset;
 
-		public override Position GetDefaultPosition() {
-			return PlayerDrawLayers.BeforeFirstVanillaLayer;
-		}
+        private void DrawAuriZone(ILContext il)
+        {
 
-		public override bool GetDefaultVisibility(PlayerDrawSet drawInfo) {
-            if (drawInfo.shadow != 0)
-                return false;
-            AuribloomPlayer modded = drawInfo.drawPlayer.GetModPlayer<AuribloomPlayer>();
-            if (modded.hasAuri)
-                return true;
-            return false;
-		}
+            try
+            {
+                ILCursor c2 = new ILCursor(il);
+                c2.GotoNext(i => i.MatchCall<Terraria.Main>("DrawInfernoRings"));
+                c2.Index++; //Move cursor to after inferno is drawn
+                c2.EmitDelegate(() =>
+                {
+                    foreach (var player in Main.ActivePlayers)
+                    {
+                        if (player.outOfRange || !player.GetModPlayer<AuribloomPlayer>().hasAuri || player.dead)
+                        {
+                            continue;
+                        }
+                        if ((player.team != Main.LocalPlayer.team || player.team == 0) && player != Main.LocalPlayer) {
+                            continue;
+                        }
+                        for (int i = 0; i < 3; i++)
+                        {
+                            float d = MathHelper.Pi * i / 6f;
+                            float auriDelta = player.GetModPlayer<AuribloomPlayer>().auriDelta;
+                            Main.spriteBatch.Draw(
+                                texture: texAsset.Value,
+                                position: player.Center - Main.screenPosition,
+                                sourceRectangle: null,
+                                color: new Color(new Vector4(0.85f + 0.15f * (float)Math.Sin(0.1f * auriDelta))),
+                                rotation: d + 0.075f * auriDelta + 0.025f * (float)Math.Sin(d * 3f + auriDelta),
+                                origin: texAsset.Size() / 2,
+                                scale: 1f + 0.05f * (float)Math.Sin(auriDelta * 0.1f),
+                                effects: SpriteEffects.None,
+                                layerDepth: 0f
+                            );
+                        }
+                    }
+                });
+            }
+            catch (Exception)
+            {
+                // If there are any failures with the IL editing, this method will dump the IL to Logs/ILDumps/{Mod Name}/{Method Name}.txt
+                MonoModHooks.DumpIL(ModContent.GetInstance<SuperfluityTwo>(), il);
 
-		protected override void Draw(ref PlayerDrawSet drawInfo) {
-			if (drawInfo.drawPlayer.dead) {
-				return;
-			}
-            Texture2D texture2d = texAsset.Value;
-            DrawData glow = new DrawData(
-                texture: texture2d,
-                position: new Vector2((int)(drawInfo.Center.X - Main.screenPosition.X), (int)(drawInfo.Center.Y - Main.screenPosition.Y)),
-                sourceRect: new Rectangle(0, 0, texture2d.Width, texture2d.Height),
-                color: new Color(1f, 1f, 1f, 1f),
-                rotation: 0,
-                origin: new Vector2(texture2d.Width / 2, texture2d.Height / 2),
-                scale: 1f,
-                effect: SpriteEffects.None
-            );
-            drawInfo.DrawDataCache.Add(glow);
-		}
+                // If the mod cannot run without the IL hook, throw an exception instead. The exception will call DumpIL internally
+                // throw new ILPatchFailureException(ModContent.GetInstance<ExampleMod>(), il, e);
+            }
+        }
     }
 }